requirements.txt
tensorflow-gpu==1.15
keras==2.1.0
tensorflow-model-optimization==0.1.3
cleverhans==3.0.1
numpy==1.16.3
scipy==1.1.0
